ai_engine.py
from google.cloud import aiplatform
import vertexai
import os
from PIL import Image  # Using Pillow for robust image handling in Simulation Mode

# -------------------------------------------------------------------------
# NOTE: The 'VirtualTryOnModel' is currently in specific Private Preview.
# The class below is a placeholder that simulates the API structure so 
# we can validate your architecture (Telegram Bot -> AI Engine -> Payment).
# Once your project is allow-listed, this class will be available in the SDK.
# -------------------------------------------------------------------------
class VirtualTryOnModel:
    @classmethod
    def from_pretrained(cls, model_name):
        print(f"🔹 [AI Engine] Loading Model: {model_name}")
        return cls()

    def drape(self, base_person_image, product_images, clothing_type):
        print("🔹 [AI Engine] Processing Drape Request...")
        print(f"   - Clothing Type: {clothing_type}")
        print(f"   - Product Images: {len(product_images)} items")
        print("   - (Simulation: Returning base image as result)")
        
        # In this simulation, we just return the base image of the model (Ananya/Zara)
        # to prove the pipeline works.
        class MockResponse:
            def __init__(self, img): self.generated_image = img
        return MockResponse(base_person_image)

def generate_drape(user_data, model_choice):
    # Set up credentials and project configuration
    if os.path.exists("service_account.json"):
        os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "service_account.json"
    
    # Initialize Vertex AI (Standard init)
    aiplatform.init(project="etoffee-engine", location="us-central1")
    vertexai.init(project="etoffee-engine", location="us-central1")

    # Load our professional base models
    model_paths = {
        "Ananya": "assets/models/ananya.png",
        "Zara": "assets/models/zara.png"
    }
    
    print(f"🔹 [AI Engine] Selected Model: {model_choice}")
    
    # Ensure the model file actually exists
    if not os.path.exists(model_paths[model_choice]):
        raise FileNotFoundError(f"Base model image not found: {model_paths[model_choice]}")

    # Initialize Vertex AI Try-On
    vto_model = VirtualTryOnModel.from_pretrained("virtual-try-on-preview-08-04")
    
    # Send all 3 shopkeeper photos as reference
    # Using PIL.Image.open() is safer for now than vertexai.Image.load_from_file()
    # because we are in simulation mode and want to avoid strict API validation errors.
    response = vto_model.drape(
        base_person_image=Image.open(model_paths[model_choice]),
        product_images=[
            Image.open(user_data['body_path']),
            Image.open(user_data['border_path']),
            Image.open(user_data['pallu_path'])
        ],
        clothing_type="one_piece" 
    )
    
    return response.generated_image
