requirements.txt
python-telegram-bot
google-cloud-aiplatform
vertexai
Pillow
nest_asyncio
python-dotenv
